Added the landing page
I added the landing page for the app in the form of a website. I also edited round corners of the quick button at the donation view, so it looks identical with the confirm button. The landing page needs to be responsive in terms I want to show it into various scales of the screen.